Add Voxtral CLI entrypoint and TTS benchmark
