+++ crates/voice-cli/src/bin/voxtral.rs
@@ -1,0 +1,1 @@
+include!("../cli.rs");
